pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = ""
authors = ["Eugene Tymoshenko"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
sqlalchemy = "^2.0.37"
alembic = "^1.14.1"
asyncpg = "^0.30.0"
libgravatar = "^1.0.4"
pydantic = "^2.10.5"
pydantic-settings = "^2.7.1"
slowapi = "^0.1.9"
cloudinary = "^1.42.1"
psycopg2-binary = "^2.9.10"
python-jose = "^3.3.0"
python-multipart = "^0.0.20"
uvicorn = "^0.34.0"
bcrypt = "3.2.0"
passlib = "^1.7.4"
fastapi-mail = "^1.4.2"
redis = "^5.2.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
pytest-cov = "^6.0.0"
httpx = "^0.28.1"
aiosqlite = "^0.20.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
sphinx = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_default_fixture_loop_scope = "function"
filterwarnings=["ignore::DeprecationWarning"]